third_party/WebKit/LayoutTests/fast/dom/Element/id-in-map.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>

<map title="map" id="firstmap">

<script>
description("Test using id in map element.");
    
shouldBeNonNull('document.getElementById("firstmap")');
</script>
</body>
</html>
